Refactored Install command and Vue app
Moved ExampleComponent.vue to the components directory.
Corrected component registration in the Vue app.
Made proper adjustments in the Install command.
