+++ src/resources/js/components/ExampleComponent.vue
